For the case where client code fetched listener from this view, just to resubmit it later
Problem is that in that case we would loose the original listener. One
solution would be to wrap the original listener with our code, but since
we do not need to listen to changes ourselves, we can delete those parts
as well and leave registering of listener solely to parent class
(AbsListView).